Converted all closure to reference value from the token object
It's important that this step be done atomically since it can either cause compilation errors or the program to wait for variables.

Case in point, for method3() you need to use the value of a. However, when you get the value of a from the dataflow queue, there isn't another value ready. So it blocks indefinitely.