Created code generation for DataflowQueues
First version that generates the DataflowQueues and wires them up as
necessary.

Right now the generated code does not fully compile because of the
exceptions on channelX.getVal() which throws an InterruptedException. I
need to figure a reasonable way to deal with this rather than just
throwing it in the method body.