Created temporary variables to hold channels
Previously the results of the channels were directly propagated to the
next closure. This created a lot more code since the closures had to
try/catch the InterrupteException from calling channel.getVal();

Also detected when an InterrupteException is already thrown on the
method and not add additional InterrupteException to the method
signature.