Check for ignored calls by consulting the call graph
We cannot just use the previous method below:

MethodReference declaredTarget= instruction.getDeclaredTarget();
if (AnalysisUtils.isLibraryClass(declaredTarget.getDeclaringClass()))

The declaredTarget is loaded by the Source classloader NOT the Extension
classloader which means that the check will fail and we won't add the
ignored method to our list.

Instead, we need to consult the callgraph to find the possible targets.
Then we query the possible targets to see which classloader loaded them.
If it is the Extension classloader then we add them to our list.

Additionally, since we are using the Java front-end we also need
to override the visitor for AstJavaInvokeInstruction in addition
to SSAInvokeInstruction.

See #51.