Add MonteCarlo test case for scalar dependencies
It seems that the naming problem is an inherent problem with the way
WALA does some optimization in the IR.

The variable iRun is said to be an input to stage1, while in reality it
is not. And thus when we create the bundle it is done wrongly.

See <https://groups.google.com/forum/?fromgroups=#!topic/wala-sourceforge-net/xfCNfG14P6o>

See #76.